feat : 해당 기수에 임시저장지원서/제출된 지원서가 존재하면 질문을 수정할 수 없도록 예외처리

+++ src/main/java/org/cotato/backend/recruit/admin/service/applicationQuestion/QuestionEditAdminService.java
@@ -3,6 +3,9 @@ package org.cotato.backend.recruit.admin.service.applicationQuestion;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.cotato.backend.recruit.admin.dto.request.applicationQuestion.QuestionUpdateRequest;
+import org.cotato.backend.recruit.admin.error.AdminErrorCode;
+import org.cotato.backend.recruit.admin.exception.AdminException;
+import org.cotato.backend.recruit.admin.service.application.ApplicationAdminService;
 import org.cotato.backend.recruit.admin.service.generation.GenerationAdminService;
 import org.cotato.backend.recruit.admin.service.question.QuestionAdminService;
 import org.cotato.backend.recruit.domain.generation.entity.Generation;
@@ -16,25 +19,28 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class QuestionEditAdminService {
-	private static final String PORTFOLIO_QUESTION_CONTENT =
-			"(선택) 추가로 제출할 포트폴리오(깃허브,블로그,노션,비핸스 등) 링크를 첨부해주세요. 업로드하실 포트폴리오 양식은 꼭 PDF로 변경 후 제출해주세요!.";
+	private static final String PORTFOLIO_QUESTION_CONTENT = "(선택) 추가로 제출할 포트폴리오(깃허브,블로그,노션,비핸스 등) 링크를 첨부해주세요. 업로드하실 포트폴리오 양식은 꼭 PDF로 변경 후 제출해주세요!.";
 	private static final Integer PORTFOIL_MAX_BYTE = -1; // 제한없음 -> -1 처리
 
 	private final GenerationAdminService generationAdminService;
 	private final QuestionAdminService questionAdminService;
+	private final ApplicationAdminService applicationAdminService;
 
 	// 전체 삭제 후 추가
 	@Transactional
-	@CacheEvict(
-			value = {"questions", "question"},
-			allEntries = true)
+	@CacheEvict(value = { "questions", "question" }, allEntries = true)
 	public void updateApplicationQuestions(QuestionUpdateRequest request) {
 		Generation generation = generationAdminService.getGenerationById(request.generationId());
 		QuestionType questionType = request.questionType();
 
-		List<Question> existingQuestions =
-				questionAdminService.getQuestionsByGenerationAndQuestionType(
-						generation, questionType);
+		boolean isApplicationExist = applicationAdminService.isApplicationExistInThisGeneration(generation.getId());
+
+		if (isApplicationExist) {
+			throw new AdminException(AdminErrorCode.QUESTION_CANNOT_UPDATE);
+		}
+
+		List<Question> existingQuestions = questionAdminService.getQuestionsByGenerationAndQuestionType(
+				generation, questionType);
 		questionAdminService.deleteAll(existingQuestions);
 
 		List<Question> newQuestions = getNewQuestions(request, generation, questionType);
@@ -47,31 +53,28 @@ public class QuestionEditAdminService {
 
 	private void insertPortfolioQuestion(
 			Integer sequence, QuestionType questionType, Generation generation) {
-		Question question =
-				Question.builder()
-						.generation(generation)
-						.sequence(sequence)
-						.content(PORTFOLIO_QUESTION_CONTENT)
-						.maxByte(PORTFOIL_MAX_BYTE)
-						.questionType(questionType)
-						.build();
+		Question question = Question.builder()
+				.generation(generation)
+				.sequence(sequence)
+				.content(PORTFOLIO_QUESTION_CONTENT)
+				.maxByte(PORTFOIL_MAX_BYTE)
+				.questionType(questionType)
+				.build();
 		questionAdminService.save(question);
 	}
 
 	private List<Question> getNewQuestions(
 			QuestionUpdateRequest request, Generation generation, QuestionType questionType) {
-		List<Question> newQuestions =
-				request.questions().stream()
-						.map(
-								element ->
-										Question.builder()
-												.generation(generation)
-												.sequence(element.sequence())
-												.content(element.content())
-												.maxByte(element.maxByte())
-												.questionType(questionType)
-												.build())
-						.toList();
+		List<Question> newQuestions = request.questions().stream()
+				.map(
+						element -> Question.builder()
+								.generation(generation)
+								.sequence(element.sequence())
+								.content(element.content())
+								.maxByte(element.maxByte())
+								.questionType(questionType)
+								.build())
+				.toList();
 		return newQuestions;
 	}
 }
